Created vector field data structure, compiles okay.
Has two structs. One struct (F_Vector) holds the vector data, and the other struct (IndexPair) contains the (currently 2D) coordinate system that is addressable also via a float pointer that returns the array of both x and y.

Applicable overrides were done to ensure hash mapping and comparison operations worked correctly.

The data structure itself is composed of a class called VectorField that contains: x and y boundary conditions, accessors for those boundary condition, a map that has a key of the index pair struct, and the F_Vector struct as the value.

Both constructors call upon the LoadDefaultVectorSet; which instantiates all vectors by default at zero. Then, the sparce vector field can be filled by either direct calls to the provided map reference, or can be instantiated by the constructor that takes in a const reference to a set, where in order, the map will be filed with the values provided.

Essentially, this allows for the maintaining of a default sparce matrix, and for changes to both the entire field or specific elements to be done accordingly.