Converted AdvectKernel and global AdvectCuda method to float3 instead of uni-dimensional velocity pointers.
This will prove to be more scalable to when I eventually convert it to 3 dimensions, and will also reduce input parameter bloat on methods.

Potential performance gains from this switch could also be possible due to thread behaviors, but this may be only an occurence within different implementations of parallelizations in all likelihood.