Ported the vector field struct to use hash maps (unordered_map) versus the standard red and black std::map for better cache utilization.
Fixed temporarily issues with size_x_ and size_y_ not being properly assigned in the VectorField by getting the boundary size with the square root of the vector field map size. WIll be fixed later.

Revamped traversal to maintain order by switching foreach functions (as there is no longer a maintained order within the underlying data structure) to nested for loops.

Created a hash struct to allow for proper hashing of the IndexPair key in the unordered map located in the VectorField class.

Modified fluid_sim.cpp procedurally in accordance with the data structure switch.