Ported data structures and methods to run on single data attributes.
Was achieved with an axis_data.hpp/cpp class, which allows VectorField to selectively give linear data.

Changed Density to properly adhere to being uni-dimensional and being stored with no deadweight components in the F_Vector as it was previously.

May change F_Vector to float3 for better interopability in the future possibly.