Changed the vector field implementation to a pointer array of AxisData as the main means of map storage to eliminate the need for constraining and repacking overhead and methods.
Got rid of RepackFromConstrained and DataConstrained from vector_field.hpp/cpp as a result of this optimization.

Removed F_Vector due to the implementation becoming unnecessary as a result of the new change.

Removed cuda_sim_helpers and moved the inline method to the fluid_sim_cuda CUDA header file.

Changed method definitions in the file to accomodate for the unidimensional shift from using F_Vectors; due to multiple problems pertaining due to CUDA host to device host complications with addresses.

Will likely completely get rid of templates due to the lack of multiple parameters required for HashMaps anymore on CUDA.