Fixed DeviceTransfer method in CUDAMap.
Bug caused invalid argument errors in CUDA (error code 1) due to the device pointer equivalent of the CUDAMap resetting itself, and as there was no subsequent cudaMalloc call, caused it to become a nullptr.

Fix was adding a conditional to where it will set the pointer to device_alloc_ once it is initialized (variable si toggled), otherwise, it returns false.