Added, tweaked, and improved documentation and formatting for the documentation. Added it specifically in fluid_sim.hpp.
Added AllocateDeviceData and ReallocateHostData to shift allocation from the simulation function helper methods to the start of the fluid simulation initialization and the end of each timestep respectively.

The device pointer in the CUDAMap was made public in order to be accessible to the simulation methods.

This allows for the optimization of only device pointers are being interfaced with until the end of each timestep, and the only time it is copied back to the host is when it needs to be outputted to VDB; significantly decreasing memory copies and host to device memory calls.