Adding weights to IndexPairs in order to establish LBM support further.
Secondary weights are stored as constant floats currently as approximations of 1/18 and 1/36 respectively. The default weight is 1/3, as that is the weight value for the midpoint.

As mentioned before, this is to be in accordance with the implementation of the D3Q19 lattice structure for Lattice Boltzmann implementations in 3D.