Got rid of extraneous default constructor in VertexData (mesh_manager.hpp).
Fixed issue regarding the implementation of the initialization of the index buffer. This was due to CopyBuffer being erroneously fixed to the MeshContainer vertices data. This was fixed by generalizing it to accept a const void* that can also contain the indices. Changed the vkCmdDraw call back to vkCmdDrawIndexed.

Fixed oversight that caused the Address for Vulkan Render Pass log output to not output the stream right after the addition of the message, which caused unnecessary duplicates of the message to be displayed in the log file.

Fixed oversight regarding not having the recreate swapchain function not reconstructing the depth images upon calling the re-creation method.

Added parameter in rasterizer.cuh to allow for the input of a different depth format when creating the depth attachment for the render pass.

Fixed typo in mesh_viewport.hpp that caused the fixed camera vertical FOV to be too wide (700 versus appropriate default 70).

Moved framebuffer width and height initialization out of loop, due to it being unnecessary for the value to be re-instated in the swap_chain_manager.cuh CreateSwapchainFrameBuffers method.

Changed usage_flags in the ImageHelper method InitializeImage to a const reference from const.