Created new project for meshing, and implementing a CUDA matrix class that has built-in methods for getting the determinant, transpose, adjoint matrix, and the inverse matrix.
Got camera system to work; behavior WIP. Now works along with the push constants as a buffer. Added W, A, S, and D controls. F is the default key for resetting the camera view. The logic for this is located in the MeshViewport method ManipulateCamera.

Added CreateBufferCross method in the buffer_helpers file that is for the staging and pushing of CPU and GPU accessible memory in Vulkan. Should refactor other staging pattern users (index buffer, push constants) to this in the future.