Added cross CUDA error checking methods in handler_methods.hpp/cpp that allow for outputting errors from device and host interopably (const char pointers as opposed to std::string).
Added a group allocation method for deleting matrices collectively.

Added an extra warning if host locality on the matrix is enabled but there is a device transfer called initiated.

Added a PrintMatrix method that takes in a default nullptr label that is null safe. The const char pointer is for an optional label that can be given at the beginning of the matrix print call.

Localized the randomization populate method into the matrix object definition, and made it explicitly host in both name and macro to avoid confusion pertaining to value initialization on device.