Separated the remaining methods from kernel.cu to the Matrix definition and weighted_least_squares.cu.
Moved the WLS logic to WeightedLeastSquares, which takes in a Matrix pointer. To reflect this, the diagonal matrix is now not fixed to a 4x4, but calculates the maximum side length of the input matrix for the matrix dimensions.

Created mpm.cu and mpm_kernel.cu to define mpm.cuh.

Created vector_cross.cuh (which is a WIP rip from the Renderer solution).

Made a global allocation method for the grid to populate the double pointers with particles and cells.

SimulateGPU now takes in a Grid pointer for better interopability, as opposed to a fixed default.

Grid also now takes in a default 4.0f particles per dim (density) setting, that also checks for 0 and negative values (as it is a float), and will default it to 4.0f if invalid, as well as warning the user of the invalid input.