Very good with progressing to successful preliminary CUDA interop.
Cross-OS support between Windows and Linux is implemented, but will have to be validated to confirm if there are any equivalent dependencies that are required in some contexts.

Created main header file cuda_interop_helper.cuh that manages the interop between memory, semaphores, and the simulation allocation logic.

Added some iterative/generalized methods in the Clean() method especially. The memory handles are stored in a struct called CrossMemoryHandle that is in a vector in the main class CudaInterop. This is so future operations that may require memory from the other channels will not have to be modified non-trivially if it were fixed-function.

Put in Windows Security struct. Put in an enum to have more robust OS based conditionals (versus only macros). Also handles low-level CUDA allocation permissions as well for inter-driver communication between Vulkan and CUDA.

Added cuda_interop_memory.cu and cuda_interop_semaphore.cu for the purposes of creating external memory and synchronization objects respectively that are interopable between CUDA and Vulkan. This, as mentioned before, has the method definitions in the header file cuda_interop_helper.cuh. It is able to change to the appropriate Win32 versus POSIX schemes for memory management.