Fixed WindowsSecurityAttributes destruction via removing the member from the implementation and moving it to where it is used; in ExportMemoryHandleWin32 and ExportSemaphoreHandleWin32 respectively.
Moved BeginSingleTimeCommands and EndSingleTimeCommands to vulkan_parameters.

Added additional abstraction and accessor methods regarding the VulkanParameters class. This includes VkDescriptorSets, wait and signal semaphores, and viewport rotation (the current MeshPushConstants struct for display).

Added extra safeguard for getting the CUDA device by warning if there are more than 1 CUDA devices in the system.

Changed CrossMemoryHandle syntax and order. Also added a type size parameter and changed the method definition of AddMemoryHandle to account for this extra parameter.

Added the method TotalAllocationSize to CrossMemoryHandle for readability and convienience purposes.