Doing preliminary testing regarding CUDA interop with a test method. Will hook it up soon.
Made CrossMemoryHandle contain the buffers and accompanying device memory for Vulkan. Fully gave it its own constructor and allocation methods.

Added some important destructors for both sync_structs in regards to the synchronization semaphores, as well as for the destruction of buffers and device memory, as well as the CUDA semaphores in the new method CleanSynchronization in CUDA interop.

Made two new files to inherit from cuda_interop_helper.cuh; cross_memory_handle.cu and cuda_interop_kernel.cu.

The first file is for defining the method definitions of the now expanded CrossMemoryHandle class.

The second is for the definition of the methods that are for the direct allocation or definition of the kernel call itself.