Changed ExportSemaphoreHandleWin32 to return a reference of the init params struct to fix an issue regarding memory overhangs.
Moved BulkInitialization out of TestMethod, as it will be running in the render loop.

InteropDrawFrame is a method added in the CUDA interop kernel, where it handles the CUDA synchronization with the Vulkan renderer.

Added blank constructors to AddMemoryHandle (maybe should get CrossMemoryHandle to not take in position handle, as it will be allocated later anyways).

Added log check and also added an initialization call for the CUDA interop device extensions in VulkanInitialization.cpp's LogicalDeviceInitialization method.
Made a call method in vulkan_parameters.cpp to make it nicer to read (calling from the CudaInterop helper class).