Changed the CrossMemoryHandle constructor definition to not take in a CUDAAllocationHandle or ShareableHandle, as they are allocated later prior to the instantiation of the CrossMemoryHandle.
Added a logger method to the CUDA interop header file (defined via the kernel definition file) for the outputting of the CUDA driver logger.

Added additional logger checks as well for standard CUDA operations pertaining to interop and synchronization.

Consolidated the default security descriptor get call into the MemoryAllocationProp method.

Fixed the issue with memory allocations (not using granularity in the mapping of the CUDA memory).

Added interop extensions for the Vulkan physical device.

Refactored the physical device extensions in VulkanGUIDriver in order to have it in line with the standard device interop extension change, as well as it to be more in line with C++ standards and to be easier to understand (const char double pointers were replaced with an internal vector of const char pointers).

Re-instated the method calls to ExportSemaphoreHandleWin32 (the method was temporarily unused in order to diagnose external semaphore validation errors).

Fixed the logger singleton finally to not have duplicates in regards to OutputLine with stringstream polymorphism.