Did refactors for synchronization-enabled options (using the GPU device memory as opposed to a set number of vertices and DrawIndexed at runtime).
Migrating potentially from using MeshPushConstants to StorageBuffer as it is more versatile for multi-mesh support + scene data (adding the framework for this later to be potentially handled for scene data): without having the same potential performance issues of uniform buffers- while being more flexible than Push Constants.

Have handler_classes.cpp in order to migrate the method definitions of the logger and the host randomizer to be separate for better organization.

Fixed matrix multiplication implementation in the cell_momentum.cu (column did not use the 2nd matrix column, which caused it to become out of bounds in the loop).

Added separate definition for the CrossMemoryHandle in cross_memory_handle.cuh in order to allow for the vertex storage classes to be able to have access without having to include the entire definition for CUDA interop.

Moved cross_memory_handles_ in CudaInterop to now be public; also for the above aforementioned reason of accessibility to storage classes for displaying interop mesh data.

Have preliminary method for the initialization of particle masses in the simulation (as it otherwise would not work without that). Will likely use some sort of randomization scheme in the future with cuRand.

Added some extra checks for Vulkan and CUDA status.

Added a method similar to CudaExceptionHandler for Vulkan (VulkanExceptionHandler).