Changed to particle position in the synchronization memory target from particle position (bug).
Changed to CUDA asserts (thank you fvla for the feedback), allows for check bypassing on Release builds and better optimizations due to not having to call printf potentially (compiler does not optimize for this).

Made ToString a host method in the Matrix implementation, and used it to the replace the print method (in order to be compatible with the Singleton logger for the GUI).

Refactored SimulateGrid heavy, as well as fixed an oversight of it not being completed. Reduced the complexity and passthrough matrices required.

General significant improvements to clarity, removal of leftover commented code, and lots of descriptive comments for ease of understanding the flow.