Fixed Vulkan handle pass issue, fixed ImportExternalBuffer oversight where it did not include the platform specific handle information in the buffer bind information struct for the call.
Fixed interop synchronization order bug in CudaInterop method InitializeCudaInterop.

Fixed issue that pertains to memory mapping location, and it now properly maps to the pointer allocated (while also still following granularity and VA rules) that was allocated in the simulation code.

Fixed implementation issue in AdvectParticles.

General minor refactors in the simulation methods for cleanliness- as well as extra commenting was done.