Massive inline overhaul of both the Matrix and the Vector implementation. Also inlining was done on the IndexPair.
Did heavy spring cleaning on the IMGUI/Vulkan early initialization side.

Made most of the Vulkan initialization helper class instantiations private.

Fixed matrix momentum and made it coinside with the use of templates and array allocation instead of dynamic allocation for significantly better performance and use.

General improvements in ergonomics in especially the VulkanGUIDriver, but also particularly the Swapchain manager.