Did some extra clamping for the boundary conditions for the advection.
Inlined almost all of the Grid functions.

Fixed ordering bug in the simulation loop that has the dim3 for threads and blocks not initialized before SetValue was ran. (Likely due to oversights and a workaround of present variable scoping issues.

Added extra methods to Vector3D and re-organized them.

Fixed Clamp to work (had to use a different implementation, not sure why it did not work before).

Modified how mass randomness is dealt with.

Changed compute version to compute86,sm86 (thank you fvla for catching this).